used util container composer to spin up laravel pro
docker-compose run --rm composer create-project --prefer-dist laravel/laravel .

then added depends_on to server so that running docker-compose -d
--build server would spin up all 3 containers and the --build just
checks if anything changed in our dockerfiles if not it's cached so just
a nice additonal check if the was aliased.

with all of that we can hit up localhost:8000 and see the default
laravel starter running. and if we update or in my case add a h1 and
save. a reload reflects those changes on localhost:8000